feat: Make it possible to select availability zones
This introduces a variable that allows the selection of availability
zones. This is useful when some instance types, e.g. m5zn.metal
are only available in certain availability zones.
If no value is specified the default behaviour stays the same
and all availability zones in the region are considered.

Signed-off-by: Frederic Giloux <frederic.giloux@isovalent.com>